refs #304 and #322
This is the initial commit of the FindXXX files for Amanzi
The NetCDF has been tested throughly and will be the template
for TPL Exodus, MOAB, Memsis, CGNS, XDMF.
The FIndHDF5 file is a modified module from the cmake 2.8.3
distribution. It will be altered to accomidate Amanzi build
requirements. IT IS NOT FUNCTIONAL at this time
The function files SelectSearchPath and PrintVariable are
common function tools that will be used extensively in the
Amanzi CMakeLists.txt files, module files and other include
files.