refs #322
Added exodus subdirectory to build
Reorg'd the CMakeList.txt file. Moved binaries that were not
defined as tests from the BUILD_TESTS if-block to the main make
block. Changed the test_files directory to EXO_BINARY_DIR. The
previous definition was creating exodus/test in the build directory
and this was a problem for out of source builds.
Since the libexodus target now has the full Teuchos and Epetra
libraries, adding Boost_*_LIBRARY at the link time for the test
binaries generated warnings on Mac OSX. Can remove Boost_*_LIBRARY
to get rid of the warnings, but may need these on other systems. Will
leave in link line for now and created a ticket refs #333 to track.