Removed ENABLE_HDF5 option from output library, refs #460
We are moving away from CGNS being the default output format and HDF5 be optional, to HDF5 being the default and required.  As such, removed the ENABLE_HDF5 option tags in the output CMakeLists.txt.  The tag was causing trouble with other libraries using the HDF5 output without enabling the options.