removed HEX element assumption, implemented mixed element mesh.
Modified how the element connectivity is written out and corresponding Xdmf output to remove the HEX element assumption and allow for mixed meshes.  The connectivity table now takes the form of a single vector with elem_typeID and elem_conn for each element.  2D polygons are now handled as well.

A single head visualization file is now written out.  ParaView now reads the VisIt style file so I consolidated to a single file.