refs #356
Update the HDF5 find module to handle HDF5 installations built with
CMake. The biggest change is the HDF5_LIBRARIES and HDF5_<language>_LIBRARY
variables now return CMake targets instead of file names on disk.
This makes the FindHDF5 module consistent with the CMake configuration
file created by the HDF5 build system. These imported targets
have the dependencies defined in the target properties. See the CMake
documentation for more information. The HDF5 project is moving
to a CMake build system but it is not complete and this file will need
to change as that capability matures.

Since HDF5 is a dependency for NetCDF, this change required i
changes to all those find modules who need NetCDF directly or need ExodusII.
These new find modules also return imported targets in their *_LIBRARY
and *_LIBRARIES variables.

The CMake community is debating best use practices for exteranl package
find modules. We have adopted the import target approach in the hope that
this will provide a more flexible build infastructure instead of the
AddPackageDependency function, which is string handling and requires the
find_package calls to occur in a particular order.