Updated FindMOAB
Return imported target for MOAB_LIBRARY and MOAB_LIBRARIES
Removed the if-test for NetCDF and HDF5 dependence, will assume
that both are required.

Added MOAB target to Amanzi build target in SuperBuild

Removed the HDF5_LIBRARIES in mesh_moab target_link definition.
Is not required now that MOAB is an imported target.

refs #665, #537