refs #647
Add ENABLE_HYPRE as a build option. Default is OFF.

If ENABLE_HYPRE is set to ON, search for the Ifpack in
Trilinos if IFpack is found AND been built with HYPRE, then
Ifpack_ENABLE_HYPRE is set to TRUE and the HYPRE APIs in
flow are activated.

Added new macro TRILINOS_PACKAGE_ENABLED_TPLS that searches for
enabled TPLs in each Trilinos package.

Compressed the Trilinos package search to improve code logic layout.
ML and NOX are only searched if ENABLE_Unstructured is TRUE.
Required packages that are not found throw an error. Optional
packages that are not found send warning messages and are deactivated
but do not stop the configuration.