refs #881
ExodusII 5.22 with NetCDF 4.2.1.1 builds results in errors in the exodus reader

The test exodus_reader fails with this combination if: Intel is the compiler,
or Amanzi is built with SuperBuild on RHEL6 64-bit.

For now, we have an option ENABLE_EXODUS498 (default value FALSE) to switch
between the older 4.98 version and newer versions. Option will be removed once these
issues are resolved. Only to be used for testing.