major rework of PKs including addition of pressure-head formulated overland flow
This new overland flow uses p_surf as the primary variable on cells,
where

h = { (p_surf - p_atm)/rho g  if p_surf > p_atm, 0 otherwise

is the ponded depth.

This has the advantage of tracking the subsurface's lambda on the
surface, making BCs for the coupling trivial (lambda = p_surf on the
subsurface, and Q_sub->surf = flux through surface face as a source
term for the surface).  The downside is it makes surface faces and
preconditioning ugly.  Needs much testing.