untested full Jacobian for surface flow
This is a bit ugly... in order to implement the Jacobian, the matrix
needs to know the upwinding.  Right now we are limited to the overland
flow case, where K_abs = 1 and upwinding is done by potential
difference with overlap.  This hacky code should get moved into the
upwinding class, and then the matrix could handle all upwinding?  Not
exactly sure of the right design here...