Fixed a bug in WhetStone. The consistency condition did not use the orientation of mesh faces in the formation of matrix N. This bug was in the routine that is not used in Amanzi yet.
Restructured the header file mfd3d.hh to reflect the internal
logic of the construction of the MFD methods in WhetStone.

Added a new method for calculating the inverse on mass matrix.
This method incorporates geometric information (face areas)
in the construction of the stabilization term. The new method
will eventually replace the default MFD method. refs #970