First plot generated during the build from an Amanzi run.
Huge progress, but for now this is sitting directly under
the user_guide directory. We would really like this to live
in the Tutorial section under examples, but Sphinx has some
limitations we'll need to work around.

Note, this needs the PYTHONPATH to include tools/amanzi_xml,
and AMANZI_INSTALL_DIR to be set so it can run amanzi.