Updated build system to support Alqumia/PFLOTRAN (refs #1079).
This update add "FindPackage" modules for Alqumia and PFLOTRAN, and
these set the related ALQUIMIA and PFLOTRAN variables needed in the
CMakeLists.txt file of the chemistry_pk.  So this file has been
simplified.

To provide the top-level installation directory to these FindPackage
modules, the template ''amanzi-tpl-config.cmake.in'' was updated
with entries for Alqumia, PFLOTRAN, and XERCES.