Switched html themes and removed conflicting extension (closes #1100, #1101).
The default.css from matplotlib appears to have motivated a more complete
implementation in the sphinxdoc theme.  For now we'll use this as it does
center figures properly.

The matplotlib extension to render mathematics seems have conflicted
as opposed to augmented MathJax.  The motivation for it appears to be to
help systems without a TeX installation, but we have TeX so no problems
here.