Added prototype python script to collect User Guide content (refs #1108).
Minor updates to testing and benchmarking files to build without errors.
This does work, but you'll need a build with alquimia enable to get the
plots of the chemistry benchmarks.  The dictionary that defines the
content and its structure is in collect.py.  This script is invoked
as part of ''make html''.