Added ''mycase'' copying capability for developers (refs #1108).
Developers need a simple way to try out their test cases.  I added
a ''mycase'' entry to the top-level index, and a simple dictionary
example for the key ''mycase''.  I added option handling to the
collect.py script and the ability to pass these to it on the make
commandline,e.g., to build ''mycase'' use

  make AMANZI_SECTION_OPTIONS="--mycase" html

Note this will only build the intro/background material along with
your test case.  To build the whole User Guide

  make AMANZI_SECTION_OPTIONS="--full-guide" html