Added scripts and analytic solutions to grid-aligned case (refs #1144).
This update includes input files and scripts to run the AT123D-AT
to generate the analytic solution.  The run_amanzi.py script was
modified to allow users to perform multiple runs with a small change
to a particular parameter - in this case its the order of the
advection scheme.  Plots can be made for any of the selected
slices.  Need to modularize of figure out how to pass options
from sphinx so we can generate three separate plots for the
documentation.