Added custom sphinx extension for tutorial discussion (refs #1111).
This update adds two sphinx extensions, a simple ''helloworld'', and a
custom literal include for amanzi xml files.  This is good start, and
is our first custom extension to Sphinx. This extension makes it easy
to grab whole element blocks and single lines from our v2.0 XML files,
and include them in the Sphinx output. But, long term needs to
leverage a more robust parsing of the xml tree.