Improved error handling to check for valid region names. refs #1200.
Added regions list and compare_region_names method to InputTranslator.  The regions list gathers the names of all of the regions defined in the input file.  The compare_region_names method is called each time an Assigned Regions list is created to check that all of the names in the new list exist in the known regions list.  If a region name is unknown the InputTranslator gives the user an error message and exits.  The error message includes the unknown region name, the input file section, and the list of known regions to help with debugging.