Finished possible incomplete update of ObservationXML and search (refs #1268).
The last updates in amanzi_xml were from the work to close state-dev, and
so it could be that these reflect an incomplete.  I hacked the function names
in what seemed to be consistent with new content,

 * search.getElementByPath -> search.getElementByTagPath
 * return getElementByPath( ) -> return getElementByNamePath(  )

and this works now. I'm sending this back to Ethan for a final check.