Updated TPLs version, and added conditional on boost foo (refs #1348).
The boost foo that Jeff added for GNU compilers on the Mac may only
work with OSX 10.9 and newer.  At least the -arch and -stdlib flags
are not supported by the Mac ports (4.8.2) GNU compilers that I have.
So I wrapped these options in a version test.  Probably needs more
fine tuning as more people test on their Macs.