Added an optional shift of an argument to standard math functions. This simplifies construction of many complex functions.
Renamed test flow_richards_newton_tpfa for consistency with
other tpfa tests.

Added a submodel to all source factories in flow and transport.
The source could be given either as a rate of change function or
a time-integrand of this rate function. Updated documentation.

Added translation of a new solute source option
"Source: Diffusion Dominated Release Model" as discussed with David.
Added a test XML file to this ticket demonstrating this capability. closes #1292