Uncommented enable_testing() to get tests with build (refs #1393).
Not sure why enable_testing() was commented out, but more importantly
it doesn't behave as I think it should.  It appeared as if all the
tests were built, but just not registered with CTest.  I think what
we need are some options that would allow us to "enable" the different
types of testing.  These would wrap the appropriate includes and
make it easy to have just the low-level tests connected by default.