Fixed an initialization bug after restart: there is no need to recalculate flux when it is provided by the checkpoint file. Recalculation did not use upwind of relative perm and gave bad initial fluxes.
Improved estimate of the relative perm during the projection step. A new strategy
is based on approximate solution of equation k(a) = k0 / (1-a) where a is
parameter in definition of the boundary pressure \lambda = p - a rho g \delta x,
and k0 is the relative perm value for a=0.