Set default install directories based on `pwd` (refs #1586).
I adjusted the default names so that without any options the following paths are used:

{{{
Directories:
    prefix                 =
    amanzi_install_prefix  =/ascem/amanzi/builds/bootstrap-test/install/amanzi
    amanzi_build_dir       =/ascem/amanzi/builds/bootstrap-test/build/amanzi
    tpl_install_prefix     =/ascem/amanzi/builds/bootstrap-test/install/tpls
    tpl_build_dir          =/ascem/amanzi/builds/bootstrap-test/build/tpls
    tpl_download_dir       =/ascem/amanzi/builds/bootstrap-test/build/tpls/Downloads
}}}

and if you set "prefix" it affects only the "install" paths (as before).

I also fixed the order of argument processing so that the paths and variables
printed by --print reflect the options given on the command line.