Add parameters to the native spec into the list of intial conditions for fields.
 <Parameter name="write checkpoint" type="bool" value="false" />
 <Parameter name="write vis" type="bool" value="false" />

These parameters define whether the field will be dumped in checkpoint
or vis files. By default these parameters are set to true. In the case
of checkpoints: if the field is not dumped into restart(checkpoint) file, after
restart the field will be initialized from input file data (not
restart). Necessary to add this feature into documentation. Ticket is created.