Added support for Belos' linear solvers
In AmanziBelosMVWrapper, I wrap all the vector types it looks like
you want to use with a solver.  (Basically, if I tried to compile,
and Belos complained that it didn't know what TreeVector was, I wrapped
it.)  Important note: while you guys defined most of the functions Belos
needs, like matvec, axpy, etc, it also wants to know the length of your
vector.  (I honestly don't remember why it wants that.  I think it's mostly
related to orthogonalization.)  It didn't look like that was defined,
so I wrote that function for all reasonable vector types; please modify
that function if the number it returns is incorrect.  I also wrapped your
operators in AmanziBelosOPWrapper so that Belos knows what to do with them.
LinearOperatorBelosGMRES provides an interface to Belos' GMRES that looks
the same as your Amanzi interface, and LinearOperatorFactory has an option
to return a Belos' GMRES solver.   Build_Trilinos.cmake now includes Belos
as one of the Trilinos packages to install.

NOTE: None of the tests currently use Belos' GMRES.  I could not find where
the tolerance is set, and due to the StatusTest (convergence test) issues
previously discussed, Belos' GMRES will never converge to the tolerance you
request.