Since the default plot and derive variables for AmanziS were set in the v1->v2 translator, and since that is no longer to be called, I have pulled that code into the setup code where state and derive variables for AmanziS are managed.
Note that there is still considerable confusion about assumptions made by Amanzi, in terms
of what is written to the vis files, and how the fields are named.  Worse, there is confusion
about which phases, components and solutes are allowed/expected at setup.  The original v1 spec
was intended to be rather general, in that an arbitrary number of phases were allowed;
in each phase an arbitrary number of components were allowed, and in each component an
arbitrary number of solutes were allowed.  Personally, I assumed that the real complexity
was going to be in how we associated components across phases, and solutes across components
and phases. The current state of things has sidestepped that completely.

Much of the original generality has been assumed away, both in the schema that is v2, and
in practice as v2 is translated to AmanziU native input.  For example, one is now to provide
a "liquid_phase" name, and this then contains a "liquid_component" name and, potentially,
"solute_components".  Moreover, the solute_component of the liquid_phase expects an
"aqueous_conc" specifier in its boundary and initial conditions.

The code to set up the relationships between phases, components and solutes in AmanziS is
therefore a bit of a mess because the modified assumptions have to be enforced somewhere
(once it is crystalized what they actually are), and this commit has not made much progress
in cleaning that part up.  AS a result, the code incorporated in this commit is very likely
to be modified again as the aforementioned assumptions are adjusted (either intentionally
and proactivity, or by fiat when new capabilities and example problems become adopted)