Adds ''inline'' keyword to resolving function in header (closes #1697).
The function AMRToAmanziLabelMap is defined in the PMAMR_Labels.H
header file, and is referenced by a method in a class in the interface
library.  But there isn't a real need for this header in the main of
converter_s.cc.  So the best solution seems to be to use the
''inline'' keyword to ensure that the function is resolved by the
linker.

There's other cleanup of the CMakeLists.txt files related to these
components that would be good (more readible, and more consistent),
but none of it is necessary to resolve this particular issue.