Disabled system call to getpwuid (closes #28).
HDF5 builds a c program (H5make_libsettings) to create a c function
(H5lib_settings.c) that reports the build settings.  I determined that
H5make_libsettings was segfaulting in the call to getpwuid. Since it
really doesn't need this info, it has the username already, the best
solution seemed to be to disable the call.  Of course, there isn't a
configure option to turn this feature off, so the cleanest solution I
found was to remove it from the list of system calls the configure
script tests for - the result is it is never enabled (disabled).