Shifted development to branch eugene/mstk-pinchouts (refs #67).
This update increases the TPLs collection version to 0.94.3,
and updates MSTK to version 3.0.0.  Since this version has
has an incompatible change to the interface of ME_Collapse,
I've added another version flag for MSTK to use to support
to support the old interface while we complete testing.

Note, mstk_mesh_parallel, is failing.