Fixes lots of dependency issues, builds against xsdk and tests run.
Most dependency issues were related to HDF5 or Alquimia and required
appropriate includes be added to the CMakeLists.txt files.  Also,
needed to fix the passing of HDF5_DIR to ASCEMIO.

So far this requires that the user clone spack, update to ensure it
builds nox, and patch trilinos, and then runs "spack install xsdk".
When building Amanzi, we get warnings about the hard coded sizes in
the netcdf header, and something about boost dependencies. Once these
are tracked down we need to fix the download of spack, and automatic
building of xsdk.