Refactored WhetStone. refs #275
The static registry of discretization methods has been created
using a non-templated factory (in factory.hh) as the example.
Removed old factory. Added parameter lists ias input for to all
discretization methods.

Created the example of using the new factory: diffusion solve that
uses edge-based DOFs. It has only academic value and should not be
a part of the base library. Relocated all edge-related code from
PDE_Diffusion to the test directory. Added edge_to_cells method to
the mesh class and provided its implementation for the MSTK framework.
This allowed to add full support of edge-based DOFs to the new schema.