Rework of how Amanzi works with communicators and mesh factory.  Refs #283 and #284
This touchs a TON of code.  It typedefs a Comm_type and then uses it
throughout the code, standardizing how Amanzi stores, passes, and uses
a communicator.

This also starts the process of cleaning up the mesh constructors and
mesh factory.  These are intertwined because of how variable the mesh
and factories are in their use of communicators -- rather than fix a
ton of code only to hack it out in the next commit, these overlap a
bit.