Refactor of MeshFactory.
Greatly simplifies the MeshFactory, removing the traits-based
implementation that was not really being used effectively.

This also changes the mesh construction process slightly, moving
things that are standard across meshes into the MeshFactory
constructor, and things that are mesh-specific into the call to
create.

This also removes the operator() interface of the mesh constructor,
preferring the create() method.

Refs #284