Enforces LayoutLeft on the output argument of MultiFunction.
This is important to make sure that a non-strided 1D view is passed
off to the virtual apply() call in Function's interface, which means
that we can assume a single View type (hopefully?).  Kokkos suggests
we should simply template the apply() method with View type, but the
problem with that is that apply() is virtual, and (at least in C++11)
we can't be both virtual and templated.