Adds a new function that does bilinear interpolation in space and linear interpolation in space, lazily.
Yes, this is trilinear interpolation, but the implementation is
different than straight-up trilinear interpolation.  The matrix values
for the bilinear interpolation in space are loaded lazily, tracking
the interval in time.  This allows for long time arrays that would not
allow loading all data at once.

The canonical need for this is raster-based precipitation data.  We
need to interpolate from the raster in x-y, and across, e.g. daily
time slices for multiple years.

Also adds some better error handling to HDF5Reader for when reading
fails.