Cleaned up plots in tracer and tritium benchmarks (refs #398).
Not sure the best approach for making these more uniform and consistent
at this point.  But to have more control of the size and margins
I create a figure of a specific size first and then grab axes object
of subplot(s).

I think that colors for the plots should be consistent, so for
now I left CrunchFlow (Magenta), changed PFLOTRAN to blue, and
left all the Amanzi results as red.