PARTIAL merge from dev/tpetra/operators
Had to revert some stuff in Find_SEACAS and XERCES because the broke
the build I was working with.

Also it looks like not everything merged over correctly, as I had
clobber some of the changes (e.g. mesh_function stuff).  So this may
need to be merged back into dev/tpetra/operators and Julien might have
to take a pass to get the merge fully done.  Also fixes some bugs
discovered by attempting to build the full example on Summit.