settling on concepts for defining host/device spaces, etc.
We must distinguish between DefaultHostMemorySpace and the host view
in a dual view for UVM -- these are not the same thing.  Also must
distinguish between things that are optimized for device only and may
not be UVM.  So there are a few more typdefs than would be simple, but
they seem to be needed.

All tests pass on summit, though some hacking of the CMakeCache was
needed (so these should get added with --arch=summit is provided to
bootstrap:

TESTS_REQUIRE_MPIEXEC:BOOL=ON
MPI_EXEC_NUMPROCS_FLAG:STRING=-n
MPI_EXEC_GLOBAL_ARGS:STRING=-g1 -a1
MPI_EXEC:FILEPATH=/sw/summit/xalt/1.2.0/bin/jsrun

Surely there is a better way to stick these in, but for now this works
as a post-configure hack.