cleans up merge from master, deals with a few lingering issues
* swaps X & Y so that Y = A^-1 X and Y = A X
* adds options related to coloring, gets this working again for
  BoomerAMG in both Operator and TreeOperator.  coloring can either be
  provided by user or can just follow the supermap.
* adds more careful checking of developer errors in calling
  Initialize, Update, Compute, Apply in Operator
* rather than adding a new constructor, constructs a flattened map for
  passing to TreeOperator's constructor.  This should get removed once
  TreeOperator gets refactored.

All tests pass.  Still need to clean up a few things, add some
documentation.  Also may remove UpdateInverse(), as it seems that the
full system should get re-constructed in cases where the symbolic
structure changed.  We currently cannot re-generate the symbolic
structure in-place (as we are creating the graph and then creating the
matrix from this static graph), so we would already have to
reconstruct the Matrix, so it seems odd to not reconstruct the
inverse.