This update fixes a minor bug in bootstrap.sh, and the OpenMP build for Hypre.
The --with-openmp option for Hypre doesn't configure the linking correctly
so I added find(OpenMP) at the CMakeLists.txt level, and use the appropriate
library to fix the linking problem.

Having access to OpenMP flags throughout seems useful so I think I'd like to
keep this there, but admittedly it seems to have some problems with intel.
So needs more testing.

With this update our common flags pick up the OpenMP command line options,
so all libraries that don't override these flags are building with OpenMP
when ENABLE_KOKKOS_OPENMP is set.