A few tweaks to the evaporation allocation scheme.
This now allocates some or all of a net sink to the subsurface when
there is no surface water at the new time and the sink is larger in
magnitude than the available water.  The surface sink becomes exactly
enough to remove all water, and the subsurface takes whatever is
left.  Note this is done in mpc_permafrost, so it only affects runs
using that MPC, and actually makes it potentially easier to use this
with non-frozen cases.

For the time being, this has not removed the M_subsurf and E_subsurf
variables from seb_evaluator, but they are always 0 and should be
removed.

To use this, the following input spec changes should be made:

1. the “surface-relative_permeability” evaluator should look like:

<ParameterList name="surface-relative_permeability" type="ParameterList">
  <Parameter name="field evaluator type" type="string" value="surface rel perm" />
  <ParameterList name="surface rel perm model" type="ParameterList">
    <Parameter name="surface rel perm model type" type="string" value="unfrozen fraction rel perm, limit one" />
    <Parameter name="unfrozen rel perm cutoff pressure [Pa]" type="double" value="0.0" />
    <Parameter name="unfrozen rel perm cutoff pressure, below [Pa]" type="double" value="-100.0" />
  </ParameterList>
</ParameterList>

2. Make sure that the “water retention evaluator” list has the
   parameter below – this got lost in a few examples (my fault)

   <Parameter name="use surface rel perm" type="bool" value="true" />

3. Assuming all other sources names are default, set the following key options:

  a. Surface flow PK list: “source key” = “surface-source_sink”
  (change from surface-mass_source)

  b. Subsurface flow PK list: “source key” = “source_sink”

Likely these will need to be updated when #83 is completed.