Alters Operators so that they use ApplyAssembled() when possible.
Currently, "when possible" means when the matrix has already been
assembled and the "shift min" option is not being used.  This can be
triggered in two cases -- either if the inverse has been used,
therefore assembling the matrix, and no update methods have been
called, or if the user calls AssembleMatrix() directly.

Also renames data_structures tests to match most of rest of Amanzi
conventions.