Renames parent_mesh in MeshColumn to col3D_mesh
This make it clearer that, while the MeshColumn is derived from this
mesh, it is not the parent.  If there is a parent of this mesh, that
is also the parent of the 1D mesh (conceptually, we want to think of
the MeshColumn as deriving directly from the 3D volume mesh, not the
column 3D mesh).