Adds (commented out) test for using MeshExtractedManifold
This capability is a Mesh-interface-only alternative to the framework
extraction constructors.  While only MSTK provides a framework
extraction constructor, this can be used on all (framework) meshes
that allow request_edges=True.

Note these tests currently fail -- it does not appear that
MeshExtractedManifold correctly implements the face_normal() method
when a cell is provided (at which point the normal returned should be
outwardly oriented with respect to the provided cell).