QOL -- debug cells now propogate from subsurface to surface
Since surface-subsurface coupling is done with error norms only on the
subsurface, it is often the case that surface debugging is necessary
to understand subsurface error.  But it is hard to go from subsurface
error cell GIDs to surface cell GIDs.  mpc_coupled_water and
mpc_permafrost now do that process -- if a debug cell or face in the
subsurface is in the top layer of cells of the subsurface, the surface
cell whose parent face is a face of that cell is added to the list of
surface cells to debug.