Adds patch to update SEACAS exodus.py wrapper
The SEACAS we use is old, and it includes an exodus.py that used to
work in python2.  Everything else we do is python3, and the SEACAS
repository include python3 wrappers.  This:

- adds a new patch to update this, old exodus.py.in script to the new
  version from the SEACAS repo for python3 (which is now required for
  ATS)
- updates the seacas-patch-step.sh.in to allow for multiple patch
  files
- registers this patch

With this change, ATS's meshing_ats scripts and watershed workflow can
use Amanzi TPL version of exodus (as long as bootstrap was run with
the --with-shared flag).