Splits MeshFramework into two classes -- MeshFramework and MeshFrameworkAlgorithms
Currently, MeshFramework supplies setters/getters, the coordinates,
and the topology, and MeshFrameworkAlgorithms supplies the algorithms
needed to calculate all other geometric quantities.

This split lets us ensure that MeshCache can use the algorithms long
after the MeshFramework has been deleted.  The algorithms can act on
not just the MeshFramework, but also the MeshCache object, using
cached coordinates/topology.

This supplies a solution to the problem of overriding algorithms in
derived classes, e.g. MeshColumn or MeshExtractedSurface, and making
sure that the MeshCache uses the overriding algorithm.  The previous
commit would have always used the default algorithm on the MeshCache,
even if the MeshFramework implemented a new algorithm.