Changes the default COMM in extracted meshes
Previously, extracted meshes were contructed, by default, with the
parent mesh's comm.  This changes that default to instead construct
with the comm provided in the factory's constructor, which is the same
for all other mesh types.

This is necessary because it allows the user to pass in MPI_COMM_SELF
or other communicators if either the MeshFactory::create() call is not
collective and/or if the meshes are intended to be constructed
serially (e.g. for column mesh extraction).

Note, the old behavior can be recovered by using the same factory to
create the parent mesh and the child mesh, or construct the child
mesh's factory with the parent mesh's communicator.

This does not effect the export option "create subcommunicator" which
uses MPI_Comm_split to create a subcommunicator of the parent's comm
with only ranks that have extracted entities.  That create() call must
be collective on the parent mesh.