Cleans up a few things to use communicators
1. PKs now create their verbose objects with the TreeVector's comm.
This is useful for domain set PKs to allow off-rank-0 comms to still
write and debug.  In particular this is useful with openmpi's
-output-filename option which will have each rank write its own file.

Note that the output of PKs on MPI_COMM_WORLD should not change.

2. Observables now get a comm.  This allows observations on vectors
that only exist on a subset of ranks.

3. Fixes bugs, race conditions, and other synchronization problems in
domain sets.

4. Adds missing check for suffixes in dependencies instead of just
keys in secondary_variable_field_evaluator.