Updates through WhetStone, @lipnikov see note below
All tests in WhetStone pass, except for those that use SingleFaceMesh,
which has not yet been refactored.  Those all crash with the expected
error, and are expected to pass when SingleFaceMesh is done.  Note
that currently these are all ifdef'ed out with the preprocessor

ifdef SINGLE_FACE_MESH

so either set that or search for that to remove.