First attempt at updating PK interface to a single state + tags
This is require to take ATS to a single state model.  Note this also
moves a few helper functions OUT of PK.hh and INTO PK_Physical.hh.
These helper functions assumed that the PK was a "leaf" PK in the
sense that it was not an MPC, so the natural home for these was in the
leaf PK class, PK_Physical.