Minor change (and added test) for observations with communicators
Previously, we have always assumed that observations are made on
fields that are defined on MPI_COMM_WORLD.  ISM models need to observe
on column meshes that are defined on MPI_COMM_SELF.  This adds a comm
to an UnstructedObservation and Observable, defaulting to
MPI_COMM_WORLD (referred to be the input spec option "domain", which
corresponds to a mesh from which we get the comm).

This results in the following:

1. All Observations exist on all ranks of MPI_COMM_WORLD

2. If the comm is not valid, the observable does nothing.  Note that
rank 0 of the comm always does all writing of files, so this is a
nonissue.

3. If the comm IS valid, the observeable must participate in the
observation's communication (e.g. MaxAll(), SumAll(), etc), even
though the variable being observed MAY NOT EXIST on that
rank (e.g. observe on MPI_COMM_WORLD, so write on rank 0, but the
variable exists on column:N domain, which is on rank 1).

This also adds two tests to confirm this works into
state_observations, and passes all existing tests and does not
change input spec for existing runs (the default is MPI_COMM_WORLD).