Adds check on VisualizationDomainSets to ensure same variables are visualized across all ranks
Since DomainSets lift variables to visualize on the shared mesh from
which they were constructed, while observations can be done on
individual domains in a domain set, there can be cases where there is
a variable that exists on a single domain (e.g. column:0) but not on
ALL column:*.  The result of that is that the list of "lifted" vectors
is different on different ranks, and VisDomainSet hangs.

Figuring out which vectors exist on all ranks is not trivial --
effectively we have to take the intersection (across ranks) of lists
of variable names.  This is done through a new utility StringReducer
which both does MPI_Allreduce() on strings and provides a nicer
interface to do this intersection.  Tests added.