Re-designed Tensorial and Kuzmin limiters to fix robustness issues.
Loss of robustness occur when two consequitive planes forming a feasible
set are almost parallel. Then, internal orthogonalization leads to loss
of accuracy. New design has two steps.

Step 1: use orthogonal projection on the planes. Orthogonal projection
reduces gradient norm and hence is robust in the presence of parallel
planes.

Step 2. Verify that the limited gradient inside the feasible set. If not,
use the BJ scheme.

It would be possible to perform multiple iterations in Step 1 (e.g.
until convergence) but there is not enough data supporting this strategy.