Fixed bug in fracture-matrix diffusion coefficient
The fracture-matrix diffusion coefficient is the result of a FV
scheme used to approximate D dC/dz where z is the normal to fracture
direction and D is the matrix diffusion coefficient. This coefficient
could be scled by porosity and tortuosite depending on the application.
This fix results to the 1st-order convergence rate wrt mesh refinement.