Support of time variable saturation in chemistry.
Chemistry/Amanzi uses the most recent saturation. This is not entirely correct when
the saturation changes in time and reactive transport subcycles w.r.t. flow. Now,
saturation which goes into a chemical beaker is interpolated between two time states.
Alqumia's implemnetation is more invoved, see #673, and not addressed here. It is
somewhat relevant to refs #673