Compute reconstructed discharghe on a single cell
This required the computation of the gradient as well
on that single cell, hence the inclusion of a few new
functions in ReconstructionCellLinear. The thing is
that for the computation of the gradient from the junction,
we do not have all the necessary information since the
neighboring cells are all pipe cells so SW water depth
is not defined there. Hence for those cells I am using
directly the pipe water depth.