Added translation of model=constant.
The question is should we always use the keywork "model" or
allow for two special cases when the model is recovered from
attributes such as "value" and "formula"?

I am leaning toward having always the model keyword which would
allow us to add checks for the number of attributes and their names.