Added a new aperture evalutor for Darcy flow.
The new evaluator is based on the linearization a(p) = a0 + c (p - p0)
where a0 and p0 are the reference aperture and pressure, respectively,
and c is the compliance. The reference pressure is initialized at time 0
to the provided pressure field. The reference aperture is setup in the
converter using the provided apreture value.

Removed unused prev_aperture field.