Moving to a cleaner desing for primary and secondary evaluators.
Primary evaluators have to be created in PK constructors to resolve
uncertainty when calling RequireEvaluator() in other PKs. For example,
TP0: flow calculation and TP1: energy transport with fixed flow.

For the same reason, secondary evaluators (e.g. aperture in the
above example) must be placed in evaluators PL. This could be done in
the converter; hence, related models should be clearly indicated in the
high-level specs, e.g. by using keyword "model".