Added fracture to mechanics PK.
Refactored algorithms for counting the number of duplicated nodes.
Note: It is possible that a distributed parent mesh has a node on two or
more ranks, but the derived mesh has this node only on one rank.
This is due to definition of entities (say, nodes) in a distributed mesh
as the closure of master and ghost cells. A node on a parent mesh has
a "cell" support, but the same node on a derived mesh may have to "cell"
support.