Fixes bug in FV discretization
This actually fixes two bugs:

1. The Jacobian term in FV for Dirichlet BCs did not respect the
upwinding scheme, so terms were included even if upwinding meant that
the kr was not a function of the cell pressure (only a function of the
boundary condition pressure).

2. HasComponent("boundary_face") returned False even in cases when
HasComponent("face") was true, and therefore the Vandelay map could be
used to create a boundary_face on the fly.  This incorrectly resulted
in several evaluators not computing on boundary_faces when
dependencies were on faces.