Reverts semantic change of HasComponent("boundary_face")
The question is whether HasComponent("boundary_face") is valid if the
vector space has "face" (and not "boundary_face").  If it has "face"
but not "boundary_face", ViewComponent("boundary_face") is valid for a
read-only view, but doesn't make much sense for a non-const view.  So
the question is, do more parts of the code rely on
HasComponent("boundary_face") to query whether to access a view for
writing or for reading?

For now, defaulting to the status quo and fixing places that expect it
for reading.